--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -55,6 +55,8 @@ Suggests:
     testthat (>= 3.1.8)
 Config/Needs/website: tidyverse/tidytemplate, xml2
 Config/testthat/edition: 3
+Config/testthat/parallel: TRUE
+Config/testthat/start-first: github-actions, release
 Encoding: UTF-8
 Language: en-US
 Roxygen: list(markdown = TRUE)