* Enable parallel testing. For me, this dropped runn time down from 30s to 11s, 10s of which is the release tests.
* Refactor release tests to avoid expensive CRAN downloads. This has equivalent test coverage but reduces test time from 10s to 4s, bringing total test time down to around 8 seconds.
* Start the slowest tests (github-actions and release) first. This brings total test time down to around 6s