@hadley 2023-10-31
Abstractly I think the turns the tests around 90°; they still cover the same stuff but in a rather different way.